[FLINK-33632] Adding custom flink mutator
